Create README.md
This is a simple command-line implementation of the classic Hangman game, developed in the C programming language. The player tries to guess a randomly selected word by suggesting letters within a limited number of attempts. With each incorrect guess, the number of remaining lives decreases. The game continues until the player either successfully guesses the entire word or runs out of lives.....

#It includes:-
1. Random word selection from a predefined list.

2. Input validation to avoid repeated guesses.

3. Tracks the number of correct and incorrect attempts.

4. Displays the current state of the word after each guess.

5. Text-based interface (runs in terminal/console).